adminApi/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
mysqlclient
channels
djangorestframework
markdown
django-filter
channels_redis
djangochannelsrestframework
